requirements.txt
matplotlib>=3.8.4
numpy>=1.26.4
opencv-python>=4.9.0.80
pandas>=2.2.2
pillow>=10.3.0
kaggle>=1.6.12
fsspec>=2024.6.0
tensorflow>=2.16.1
scikit-learn>=1.5.0
keras>=3.3.3